bump to GPLv3

+++ G2MainServer.h
@@ -7,8 +7,9 @@
  * This file is part of g2cd.
  *
  * g2cd is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version
- * 2 as published by the Free Software Foundation.
+ * it under the terms of the GNU General Public License as
+ * published by the Free Software Foundation, either version 3
+ * of the License, or (at your option) any later version.
  *
  * g2cd is distributed in the hope that it will be useful, but
  * WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -16,9 +17,8 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public
- * License along with g2cd; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston,
- * MA  02111-1307  USA
+ * License along with g2cd.
+ * If not, see <http://www.gnu.org/licenses/>.
  *
  * $Id: G2MainServer.h,v 1.16 2005/09/05 10:18:46 redbully Exp redbully $
  */
@@ -130,7 +130,7 @@ _G2MAIN_EXTRNVAR(struct
 			{
 				enum g2_qht_comp compression;
 				bool compress_internal;
-				size_t max_promille;
+				unsigned max_promille;
 			} qht;
 		} settings;
 		struct
